requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
Pillow==10.0.1
gTTS==2.4.0
numpy==1.24.3
Werkzeug==2.3.7
google-cloud-vision==3.4.4
google-auth==2.23.4
